GUI Update, More Serialization, GameUpdate
Added NetworkGameList + GameUpdate (see server commit), Command (see server commit)
Added login functionality, the client sends a login request to the server with a username and hashed (SHA256) password. If it is a match the client will receive a instance of Player from the server which also acts as the users profile, keeping track of wins and losses.
Updated the GUI to have a login screen before the main player control window opens, i.e., what the player will go into before starting a game. I had a very big problem with the GUI being a singleton and being the main method. Whenever any other class, such as GameClient(what handles the client to server requests) asked for the instance of GameWindow it would refresh the window to it’s base state. When a login request was processed the client would not update and go into the game menu but would refresh the login page to It’s startup form and not login. This was a huge pain, but the fix was removing the Singleton and making the GameWindow static. This fixed that bug but required a lot of changes to the code to ensure all current methods used the static modifier or just did not ask for the instance but the class.